Create README.md
Preface:
Defitt & FTT COIN is a crypto token based on ERC-20 that can be purchased with Ether, the native 
cryptocurrency of the Ethereum blockchain. ERC-20 is simply a standard that specifies how these 
tokens behave, so that they are compatible with other platforms like cryptocurrency exchanges.
In the following sections, we give the details of the underlying architecture that we have used to 
develop the coin and also we give a security & safety assessment report of the technology being 
used.